Trait valico::common::error::ValicoError

pub trait ValicoError: Error + Debug + Send + Any + ToJson {
    fn get_code(&self) -> &str;
    fn get_path(&self) -> &str;
    fn get_title(&self) -> &str;

    fn get_detail(&self) -> Option<&str> { ... }
}

Required Methods

fn get_code(&self) -> &str

fn get_path(&self) -> &str

fn get_title(&self) -> &str

Provided Methods

fn get_detail(&self) -> Option<&str>

Methods

impl ValicoError

fn is<T: ValicoError>(&self) -> bool

Returns true if the boxed type is the same as T

fn downcast_ref<T: ValicoError>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_ref_unchecked<T: ValicoError>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

fn downcast_mut<T: ValicoError>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_mut_unchecked<T: ValicoError>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl ValicoError

fn downcast<T: ValicoError>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

unsafe fn downcast_unchecked<T: ValicoError>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.