Struct valico::json_schema::builder::Builder

pub struct Builder {
    // some fields omitted
}

Builder provides simple DSL to build Schema. It allows you not to use strings and raw JSON manipulation. It also prevent some kinds of spelling and type errors.

Methods

impl Builder

fn new() -> Builder

fn id(&mut self, url: &str)

fn ref_(&mut self, url: &str)

fn schema(&mut self, url: &str)

fn default<T>(&mut self, default: T) where T: ToJson

fn multiple_of(&mut self, number: f64)

fn maximum(&mut self, number: f64, exclusive: bool)

fn minimum(&mut self, number: f64, exclusive: bool)

fn max_length(&mut self, number: u64)

fn min_length(&mut self, number: u64)

fn pattern(&mut self, pattern: &str)

fn format(&mut self, format: &str)

fn items_schema<F>(&mut self, build: F) where F: FnOnce(&mut Builder)

fn items_array<F>(&mut self, build: F) where F: FnOnce(&mut SchemaArray)

fn additional_items(&mut self, allow: bool)

fn additional_items_schema<F>(&mut self, build: F) where F: FnOnce(&mut Builder)

fn max_items(&mut self, number: u64)

fn min_items(&mut self, number: u64)

fn unique_items(&mut self, unique: bool)

fn max_properties(&mut self, number: u64)

fn min_properties(&mut self, number: u64)

fn required(&mut self, items: Vec<String>)

fn properties<F>(&mut self, build: F) where F: FnOnce(&mut SchemaHash)

fn pattern_properties<F>(&mut self, build: F) where F: FnOnce(&mut SchemaHash)

fn additional_properties(&mut self, allow: bool)

fn additional_properties_schema<F>(&mut self, build: F) where F: FnOnce(&mut Builder)

fn dependencies<F>(&mut self, build: F) where F: FnOnce(&mut Dependencies)

fn enum_<F>(&mut self, build: F) where F: FnOnce(&mut ArrayBuilder)

fn array(&mut self)

fn boolean(&mut self)

fn integer(&mut self)

fn number(&mut self)

fn null(&mut self)

fn object(&mut self)

fn string(&mut self)

fn type_(&mut self, type_: PrimitiveType)

fn types(&mut self, types: &[PrimitiveType])

fn all_of<F>(&mut self, build: F) where F: FnOnce(&mut SchemaArray)

fn any_of<F>(&mut self, build: F) where F: FnOnce(&mut SchemaArray)

fn one_of<F>(&mut self, build: F) where F: FnOnce(&mut SchemaArray)

fn not<F>(&mut self, build: F) where F: FnOnce(&mut Builder)

fn build<F>(build: F) -> Builder where F: FnOnce(&mut Builder)

fn into_json(self) -> Json