Trait valico::json_schema::keywords::Keyword

pub trait Keyword: Sync + Any {
    fn compile(&self, &Json, &WalkContext) -> KeywordResult;
}

Required Methods

fn compile(&self, &Json, &WalkContext) -> KeywordResult

Methods

impl Keyword

fn is<T: Keyword>(&self) -> bool

Returns true if the boxed type is the same as T

fn downcast_ref<T: Keyword>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_ref_unchecked<T: Keyword>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

fn downcast_mut<T: Keyword>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_mut_unchecked<T: Keyword>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl Keyword

fn downcast<T: Keyword>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

unsafe fn downcast_unchecked<T: Keyword>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.